Move LocaleManager to client-side
