+++ src/main/java/es/deusto/spq/client/gui/locale/LocaleMode.java
@@ -1,4 +1,4 @@
-package es.deusto.spq.server.locale;
+package es.deusto.spq.client.gui.locale;
 
 /**
  * LocaleMode sets the expected behaviour of the LocaleManager class
